CortexVision/requirements.txt
flask
flask-sock
snowflake-connector-python
pandas
Pillow
selenium
webdriver-manager
scraper
ChromeDriverManager
beautifulsoup4
requests
python-dotenv